Add static assertions for struct layout to generated C++ code.
PiperOrigin-RevId: 399627054

+++ rs_bindings_from_cc/ir.rs
@@ -70,7 +70,7 @@ pub struct Field {
     #[serde(rename(deserialize = "type"))]
     pub type_: MappedType,
     pub access: AccessSpecifier,
-    pub offset: i64,
+    pub offset: usize,
 }
 
 #[derive(Debug, PartialEq, Eq, Hash, Clone, Deserialize)]
